add hooks for handling search with URL syncing
This adds a Search form for searching the pokedex for Pokemon by name.
The search is kept in sync with the query string and vice versa. So far,
the search results are completely unstyled.

Misc improvements made to add process.env.ENV_VAR so we can have a
POKEDEX_URL.